qcacld-3.0: Remove #ifDefender FEATURE_WLAN_LFR from SME
The condition #ifDefender WLAN_FEATURE_LFR is used to include some of
the roaming code. It has been Defenderined for all the recent driver
builds. Roaming code underneath is permanent part of the driver.
Remove #ifDefender FEATURE_WLAN_LFR compiler flag from SME so that the code
will always be compiled in.

CRs-Fixed: 936683
Change-Id: I86b9e3dee6b92c89a54de6df769800c31fc3ad17