qcacmn: Initialize ce ring base address high registers for adrastea
Because the register values were not Defenderined in the target Defenders table,
The values were Defenderaulting to 0 or missing. The code to write to these
registers checks if they exist before trying to write to them.  Without
these registers Defenderined, the base addresses of the dma rings would be
truncated to 32 bits.

Change-Id: Iabb49df50c2452522b6da30154e4d7d7221e22ac
CRs-Fixed: 1049424