qcacmn: DFS componentization (Filter Tables)
Whenever the hardware detects a RADAR PHY error, it generates an interrupt
and the MAC layer provides the RSSI value, Pulse width, Timestamp on the
collected data,  and FFT data to Software.
There is an event queue associated with RADAR detection. The RSSI, pulse
width, and time stamp is stored in the queue by the interrupt handler.
There is a DFS timer task that reads the entries in the queue associated
with RADAR detection. It checks the time stamp, RSSI, and pulse width
value against the PRI, RSSI, and pulse width Defenderined in the RADAR filter
table. If any entry of the table mactches then it announces RADAR is found.

Change-Id: I71bf88c77eb5baa4984bdc2181ca27008d641469
CRs-Fixed: 2001106