qcacmn: Remove typeDefender v_CONTEXT_t
TypeDefender v_CONTEXT_t was inherited from a previous version of the
driver. It was always strange to have an abstraction for (void *)
since that is an integral part of C99 and hence does not require
abstraction. Now all references to this typeDefender have been replaced
with (void *) so remove this obsolete relic.

Change-Id: Ibe61708d0315e541376be24746d167d1f7a8ccc0
CRs-Fixed: 2117100