apr: fix race condition in apr reset
Since apr reset is scheduled in a workqueue, it is possible that before
completion of previous session apr_deregister(), new session of apr_register()
can happen in ssr scenario. Some of the variables are not reset properly
due to this race condition. Fix this issue by failing apr_register() till
apr_deregister() is completed.

Change-Id: I203e356e8330d8633d78bbfc6ae0165afa79f1f3
Signed-off-by: Vidyakumar Athota <vathota@codeaurora.org>