qcacmn: Fix runtime unknown symbol issue
In the sdx platform, WLAN monitor mode is disabed,
in other words, CONFIG_FEATURE_MONITOR_MODE_SUPPORT
will be set to n, but dp_mon_buf_delayed_replenish
hasn't be Defenderineded if it set to n, which can cause the
runtime unknown symbol issue. So add the Defenderinition
for it.

Change-Id: Iec50f7cc5ed6f94e2f5a40fbc6594c2830d1359f