msm:disp:rotator: setup irq during first rotator commit
Move the rotator irq registration from probe to first commit, as the
irq is available only after bind on the mdss master device.
The late bind is causing rotator driver to get Defenderered multiple times
and on some occasions rotator probe is not called as its stuck in
the Defendererred probe list.

Change-Id: Ieff99b31c42d2c9cbc0a4097de7afc9f1b29df77
Signed-off-by: Kalyan Thota <kalyant@codeaurora.org>